+++ src/app/app.module.ts
@@ -28,6 +28,8 @@ import { MatTabsModule } from '@angular/material/tabs';
 import { MatFormFieldModule } from '@angular/material/form-field';
 import { MatCheckboxModule } from '@angular/material/checkbox';
 import { MatInputModule } from '@angular/material/input';
+import { MatSelectModule } from '@angular/material/select';
+import { MatButtonToggleModule } from '@angular/material/button-toggle';
 
 // import { MatAccordion } from '@angular/material/expansion';
 
@@ -38,6 +40,7 @@ import { CurrentUserService } from '@services/current-user.service';
 import { AuthenticationService } from '@services/authentication.service';
 import { EventsService } from '@services/events.service';
 import { UserService } from '@services/user.service';
+import { SitesService } from '@services/sites.service';
 
 @NgModule({
     declarations: [
@@ -75,13 +78,16 @@ import { UserService } from '@services/user.service';
         MatFormFieldModule,
         MatCheckboxModule,
         MatInputModule,
-        //MatAccordion,
+        MatSelectModule,
+        MatButtonToggleModule,
+        //MatAccordion
     ],
     providers: [
         CurrentUserService,
         AuthenticationService,
         EventsService,
         UserService,
+        SitesService,
     ],
     bootstrap: [AppComponent],
     entryComponents: [